feat(sdk): code refactoring
